requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.51.2
httpx==0.27.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
psutil==5.9.6
python-docx==0.8.11
Pillow==10.1.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
pydantic==2.5.1
secure==0.3.0
bcrypt==4.1.2
PyJWT==2.8.0 